Removed all redundant typecasts.

+++ openCVLibrary24133/src/main/java/org/opencv/core/RotatedRect.java
@@ -25,11 +25,11 @@ public class RotatedRect {
 
     public void set(double[] vals) {
         if (vals != null) {
-            center.x = vals.length > 0 ? (double) vals[0] : 0;
-            center.y = vals.length > 1 ? (double) vals[1] : 0;
-            size.width = vals.length > 2 ? (double) vals[2] : 0;
-            size.height = vals.length > 3 ? (double) vals[3] : 0;
-            angle = vals.length > 4 ? (double) vals[4] : 0;
+            center.x = vals.length > 0 ? vals[0] : 0;
+            center.y = vals.length > 1 ? vals[1] : 0;
+            size.width = vals.length > 2 ? vals[2] : 0;
+            size.height = vals.length > 3 ? vals[3] : 0;
+            angle = vals.length > 4 ? vals[4] : 0;
         } else {
             center.x = 0;
             center.x = 0;
@@ -42,8 +42,8 @@ public class RotatedRect {
     public void points(Point pt[])
     {
         double _angle = angle * Math.PI / 180.0;
-        double b = (double) Math.cos(_angle) * 0.5f;
-        double a = (double) Math.sin(_angle) * 0.5f;
+        double b = Math.cos(_angle) * 0.5f;
+        double a = Math.sin(_angle) * 0.5f;
 
         pt[0] = new Point(
                 center.x - a * size.height - b * size.width,
